FinalProjectProg1inC
This program intends to support its users in the process of managing the minutes of an organization's general meetings. The program is managed by administrators who can register new users. Each user has a unique username, a name, a password, and a type (administrator, shareholder). A minute has several pieces of information, namely its number (sequential); the date and place of the meeting to which it relates; the text of the minutes; the list of shareholders present at the respective meeting, recorded in order of arrival. Each minute has a status that can be one of the following: in writing; for subscription; concluded